implemented change req. #46 propertyNames()
in interface Accessible

    Set<String> propertyNames();

This differs from the correspondent method in java.util.Properties
since Properties.propertyNames returns an Enumeration<?> but property
names are required to be Strings. In JDK 1.6 the method
`Set<String> stringPropertyNames()` has been introduced but since
we require JDK 1.5+ support I couldn't just delegate and I had to
reimplement it. So I chosed to implement it with the proper method
signature, even though it doesn't match java.util.Properties
interface.