improved file changes watch
When two URI are pointing to the same file, example:

@Sources({
"${catalina.home}/conf/foobar.properties",
"${catalina.base}/conf/foobar.properties"
})

if ${catalina.home} == ${catalina.base} they may point to the same resource.

In this case, we are using two WatchableFile instance to check the same file, even though it is the same file.

This change is intended to recognize that the two File objects are pointing to the same instance and handle as a single WatchableFile.