Add list() to ConfigCache
ConfigCache is a great way to centralise configuration for various
parts of an application. This commit adds a list() method to the
ConfigCache class, which lists the keys for all configurations
present in the cache. This allows the entire application
configuration to be inspected (e.g. for debugging) without the need
for storing cache keys elsewhere.