Add support for deserializing nested JSON documents (using google gson)
For developers not wanting to use json, and to keep the dependency foot-print
at a minimum, the gson dependency is set to "provided".

The `LoadersManager` will check if Google's Gson parser is available on
path before attempting to register the JSON parser. (I'm happy to change this
behavior if you feel there's a better way to do this)

Usage -

```json
{
  "server": {
    "http": {
      "port": 80,
      "hostname": "localhost"
    },
    "ssh": {
      "port": 22,
      "address": "127.0.0.1",
      "alive": {
        "interval": 60
      },
      "user": "admin"
    },
    "admins": ["bob", "alice"]
  }
}
```

```java
@Config.Sources({"classpath:org/aeonbits/owner/server.test.json"})
public static interface ServerConfig extends Config, Accessible {

    @Key("server.http.port")
    int httpPort();

    @Key("server.http.hostname")
    String httpHostname();

    @Key("server.ssh.port")
    int sshPort();

    @Key("server.ssh.address")
    String sshAddress();

    @Key("server.ssh.alive.interval")
    int aliveInterval();

    @Key("server.ssh.user")
    String sshUser();

    @Key("server.admins")
    List<String> admins();
}
```